requirements.txt
Flask==2.0.2
Flask_WTF==1.0.0
matplotlib==3.5.1
numpy==1.19.5
Pillow==9.0.1
python-dotenv==0.19.2
scikit_image==0.19.1
tensorflow==2.5.0
Werkzeug==2.0.2
uvicorn
